# openapi/mobile-v1.yaml
openapi: 3.0.3
info:
  title: FusionBikes — API móvil v1
  version: 1.0.0
  description: |
    Contrato congelado para la app móvil (React Native). Fuente de verdad para ambos lados
    (app y backend) — ver docs/superpowers/plans/plan-maestro-v2.md para el plan vigente.
    La autenticación, dispositivos, notificaciones e inbox operativo de Claims están
    implementados bajo /api/v1. Los demás paths se incorporan por entregas verticales.

    Convenciones:
    - Todas las fechas en ISO 8601 UTC.
    - Todos los IDs de recurso son string (aunque el backend interno use INTEGER autoincrement).
    - El shape de error es uniforme en todos los endpoints (ver `Error` en components/schemas).
    - Paginación por cursor en listados (`cursor`/`next_cursor`), no por offset.
servers:
  - url: https://herramientas.fusionbikes.com.ar/api
    description: Producción (cuando exista)
  - url: http://localhost:3001/api
    description: Local (DISABLE_CRONS=true, nunca contra data/fusion.sqlite real)
...
security:
  - bearerAuth: []
paths:
...
  /operations/{id}:
    get:
      tags: [stock]
      summary: Consulta el resultado (posiblemente asincrónico) de una operación de stock
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operation' }
              example:
                id: "op_9f8c3e"
                tipo: "stock_adjustment"
                woo_status: "aplicado"
                ml_status: "sincronizado"
                detalle_ml: "Stock actualizado: 7"
                creado_en: "2026-08-27T10:00:00Z"
                actualizado_en: "2026-08-27T10:00:03Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /integration-notifications:
    get:
      tags: [inbox]
      summary: Lista notificaciones lógicas del backbone de integraciones
      description: Requiere el permiso notificaciones-ml y solo devuelve filas del usuario autenticado.
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        '200':
          description: Notificaciones lógicas y cursor siguiente
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/IntegrationNotification' } }
                  next_cursor: { type: string, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /conversations/{id}:
    get:
      tags: [inbox]
      summary: Lee una conversación visible y sus mensajes
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      responses:
        '200':
          description: Conversación y mensajes
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConversationView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /conversations/{id}/read:
    post:
      tags: [inbox]
      summary: Marca como leídos los ítems de una conversación
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      responses:
        '200':
          description: Resultado de la actualización
          content:
            application/json:
              schema:
                type: object
                required: [ok, updated]
                properties:
                  ok: { type: boolean }
                  updated: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /operations/{correlationId}:
    get:
      tags: [inbox]
      summary: Consulta etapas seguras de un evento por correlation_id
      parameters:
        - name: correlationId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Evento y trazabilidad operativa
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationOperation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    Unauthorized:
      description: Sin sesión válida o sesión expirada
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { ok: false, error: "No autenticado" }
    Forbidden:
      description: Autenticado pero sin permiso para esta acción/recurso
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: "sin_permiso", message: "Tu usuario no tiene permiso para esta acción" } }
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: "no_encontrado", message: "El recurso solicitado no existe" } }
...
    ServerError:
      description: Error interno no esperado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: "error_interno", message: "Ocurrió un error inesperado, reintentá" } }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: "Identificador estable, no cambia entre versiones" }
            message: { type: string, description: "Texto legible, en español, para mostrar o loguear" }
...
    Operation:
      type: object
      properties:
        id: { type: string }
        tipo: { type: string, example: "stock_adjustment" }
        woo_status: { type: string }
        ml_status:
          type: string
          enum: [sincronizando, sincronizado, sin_cambios, error, omitido]
        detalle_ml: { type: string, nullable: true }
        creado_en: { type: string, format: date-time }
        actualizado_en: { type: string, format: date-time }
...
    IntegrationNotification:
      type: object
      required: [notification_id, user_id, title, status, created_at]
      properties:
        notification_id: { type: integer }
        event_id: { type: string, nullable: true }
        inbox_id: { type: integer, nullable: true }
        user_id: { type: integer }
        title: { type: string }
        body: { type: string, nullable: true }
        deep_link: { type: string, nullable: true }
        status: { type: string, enum: [pending, sent, failed, expired] }
        created_at: { type: string, format: date-time }
        read_at: { type: string, format: date-time, nullable: true }
    ConversationView:
      type: object
      required: [conversation, messages]
      properties:
        conversation: { type: object, description: Registro interno de conversación visible al usuario }
        messages:
          type: array
          items:
            type: object
            properties:
              message_id: { type: integer }
              event_id: { type: string, nullable: true }
              direction: { type: string, enum: [inbound, outbound] }
              external_message_id: { type: string, nullable: true }
              body: { type: string, nullable: true }
              body_redacted: { type: integer, enum: [0, 1] }
              occurred_at: { type: string, format: date-time, nullable: true }
              created_at: { type: string, format: date-time }
    IntegrationOperation:
      type: object
      required: [correlation_id, event, history]
      properties:
        correlation_id: { type: string }
        event: { type: object, description: Evento canónico sin secretos ni payload privado }
        history:
          type: array
          items:
            type: object
            properties:
              stage: { type: string }
              from_status: { type: string, nullable: true }
              to_status: { type: string, nullable: true }
              error_code: { type: string, nullable: true }
              resource_id: { type: string, nullable: true }
              correlation_id: { type: string }
              retryable: { type: integer, enum: [0, 1], nullable: true }
              attempts: { type: integer, nullable: true }
              safe_message: { type: string, nullable: true }
              created_at: { type: string, format: date-time }